Revert "Add Script alias (#6237)"
This reverts commit dd65f24bb3a8932f60f62c75595d87dd9a786ebd.
