+++ src/deprecated/deprecated.js
@@ -126,9 +126,6 @@ import { basisInitialize } from '../framework/handlers/basis.js';
 import { LitShader } from '../scene/shader-lib/programs/lit-shader.js';
 import { Geometry } from '../scene/geometry/geometry.js';
 
-// ScriptType alias
-export { Script as ScriptType } from '../framework/script/script.js';
-
 // CORE
 export const LINEBATCH_WORLD = 0;
 export const LINEBATCH_OVERLAY = 1;
